requirements.txt
asgiref==3.5.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
bleach==5.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-baton==2.4.1
django-ckeditor==6.5.1
django-cors-headers==3.11.0
django-editorjs==0.2.1
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-modeladmin-reorder==0.3.1
django-quill-editor==0.1.40
django-summernote==0.8.20.0
django-tinymce==3.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
google-api-python-client==1.4.1
gunicorn==20.1.0
heroku==0.1.4
httplib2==0.22.0
idna==3.3
importlib-metadata==4.12.0
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.4.1
MarkupSafe==2.1.1
oauth2client==4.1.3
oauthlib==3.2.1
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==1.5
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
zipp==3.8.1
